+++ tiatoolbox/models/__init__.py
@@ -13,3 +13,19 @@ from tiatoolbox.models.engine.semantic_segmentor import (
     SemanticSegmentor,
     WSIStreamDataset,
 )
+
+from .architecture.hovernet import HoVerNet
+from .architecture.hovernetplus import HoVerNetPlus
+from .architecture.idars import IDaRS
+from .architecture.mapde import MapDe
+from .architecture.micronet import MicroNet
+from .architecture.nuclick import NuClick
+from .architecture.sccnn import SCCNN
+
+HoVerNet = HoVerNet
+HoVerNetPlus = HoVerNetPlus
+IDaRS = IDaRS
+MapDe = MapDe
+MicroNet = MicroNet
+NuClick = NuClick
+SCCNN = SCCNN
